# docs/swagger.yaml
basePath: /rederinghub.io/v1
definitions:
...
  request.ListNftIdsReq:
    properties:
      inscriptionIDs:
        items:
          type: string
        type: array
      payType:
        additionalProperties:
          type: string
        type: object
      price:
        type: string
      seller_address:
        type: string
      seller_ord_address:
        type: string
    type: object
...
  request.UpsertRedisRequest:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  response.JsonResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.RespondErr'
      status:
        type: boolean
    type: object
...
  response.RedisResponse:
    properties:
      value:
        type: string
    type: object
...
  response.RespondErr:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server Generative.
  title: Generative.xyz APIs
  version: 1.0.0
paths:
...
  /admin/auto-listing:
    post:
      consumes:
      - application/json
      description: Auto listing
      parameters:
      - description: ' Auto listing'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ListNftIdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: boolean
              type: object
      summary: Auto listing
      tags:
      - Admin
  /admin/redis:
    delete:
      consumes:
      - application/json
      description: Delete Redis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Redis
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get Redis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.RedisResponse'
                  type: array
              type: object
      summary: Get Redis
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Upsert Redis
      parameters:
      - description: Upsert redis key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpsertRedisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.RedisResponse'
              type: object
      summary: Upsert Redis
      tags:
      - Admin
  /admin/redis/{key}:
    delete:
      consumes:
      - application/json
      description: Delete Redis
      parameters:
      - description: Redis key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Redis
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get Redis
      parameters:
      - description: Redis key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.RedisResponse'
              type: object
      summary: Get Redis
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"